==> AF_MA_Master_RAA/10_Gene_Performance.Rmd <==
---
title: "11_Gene_performance"
output: html_document
date: "2023-02-26"
---

# 1. Loading Packages and Functions
```{r}
library(ggplot2)
library(cowplot)

source("src/utils_AF_Meta.R")
```

# 2. Main data object importing
```{r}
AF_Meta = readRDS(file = "data/AF_Meta.rds") #main object
```

```{r}
# 0. Processing object

experiments = names(AF_Meta)
names(experiments) = experiments
experiment_size = sort(unlist(lapply(AF_Meta, 
                                     function(x) ncol(x$GEX))),
                       decreasing = T)
```

```{r}
# 1. Meta-analysis

fisher_rank = run_fisher_meta_AF(meta_list = AF_Meta,
                              n_missing = length(AF_Meta) - 5)

t_matrix = get_all_limma_AF(meta_list = AF_Meta,
                         limma_column = "t")
lfc_matrix = get_all_limma_AF(meta_list = AF_Meta,
                           limma_column = "logFC")
genes = names(fisher_rank)

```

```{r}
# 2. How the performance improves as we include more and more top genes

ngenes = seq(50,10000,100)
names(ngenes) = ngenes

InPerformance = map(ngenes, function(ingenes){ # We will include 100 genes at a time til 5k
  testgenes = names(fisher_rank[1:ingenes]) # Defining the genes to be used to calculate DS
  ingenes_stats = getRisk_Stats_AF(Experiment_List = AF_Meta,
                                   limma_t_mat = t_matrix, 
                                   genes = testgenes)
  
  ingenes_stats = map(ingenes_stats, function(x){ #Calculating mean AUC of each predicted
    enframe(x[["SingleAUC"]],       #experiment
            "PredictorExperiment",
            "AUC")
  }) %>% enframe("PredictedExperiment") %>%
    unnest() 
  #%>% group_by(PredictedExperiment) %>%
  #summarise(meanAUC = mean(AUC))
  
  return(ingenes_stats)
}) %>% enframe("nin") %>% unnest() %>% mutate(nin = as.numeric(nin),
                                              PredictedExperiment = factor(PredictedExperiment,
                                                                           levels = names(experiment_size)
                                              ))

InPerformance_df = InPerformance %>% group_by(PredictedExperiment,nin) %>%
  summarise(meanAUC = mean(AUC))


# Best performance: Saturation

saveRDS(InPerformance_df, 
        file = "data/figure_objects/InPerformance_df.rds")


InPerformance_zoom_df = InPerformance_df

max_class = InPerformance_zoom_df[order(InPerformance_zoom_df$meanAUC, decreasing = TRUE), ][1, ]


sum(max_class$nin <= 500)

saveRDS(InPerformance_zoom_df, 
        file = "data/figure_objects/InPerformance_df_zoom.rds")

```

```{r}
# 3. How the performance worsens as we exclude top genes

ngenes = seq(50,10000,100)
names(ngenes) = ngenes

OutPerformance = map(ngenes, function(outgenes){ # We will include 100 genes at a time til 5k
  testgenes = names(fisher_rank[outgenes:length(fisher_rank)]) # Defining the genes to be used to calculate DS
  ingenes_stats = getRisk_Stats_AF(Experiment_List = AF_Meta,
                                   limma_t_mat = t_matrix, 
                                   genes = testgenes)
  
  ingenes_stats = map(ingenes_stats, function(x){ #Calculating mean AUC of each predicted
    enframe(x[["SingleAUC"]],       #experiment
            "PredictorExperiment",
            "AUC")
  }) %>% enframe("PredictedExperiment") %>%
    unnest() %>% group_by(PredictedExperiment) %>%
    summarise(meanAUC = mean(AUC))
  
  return(ingenes_stats)
}) %>% enframe("nout") %>% unnest() %>% 
  mutate(nout = as.numeric(nout),
         PredictedExperiment = factor(PredictedExperiment,
                                      levels = names(experiment_size)))

saveRDS(OutPerformance, 
        file = "data/figure_objects/OutPerformance_df.rds")
```

```{r}
library(ggplot2)
library(cowplot)
```

```{r}
InPerformance_df = readRDS(file = "data/figure_objects/InPerformance_df.rds")
InPerformance_zoom_df = readRDS(file = "data/figure_objects/InPerformance_df_zoom.rds")
OutPerformance_df = readRDS(file = "data/figure_objects/OutPerformance_df.rds")
```

```{r}
#Creating plot 
InPerformance_plt = ggplot(InPerformance_df, 
                           aes(x = as.numeric(nin), y = meanAUC, 
                               group = PredictedExperiment, 
                               color = PredictedExperiment)) +
  geom_line() + theme_minimal() + 
  geom_vline(xintercept = 500) + 
  theme(axis.text = element_text(size=12),
        axis.title = element_text(size=12),
        panel.grid = element_blank(),
        panel.background = element_rect(fill=NULL, colour='black',
                                        size=1)) +
  xlab("Number of top genes included in the signature")

explegend = get_legend(InPerformance_plt)
InPerformance_plt = InPerformance_plt + theme(legend.position = "none")
```

```{r}
#Creating plot 

OutPerformance_plt = ggplot(OutPerformance_df, 
                            aes(x = as.numeric(nout), y = meanAUC, 
                                group = PredictedExperiment, 
                                color = PredictedExperiment)) +
  geom_line() + theme_minimal() + 
  geom_vline(xintercept = 500) + 
  theme(axis.text = element_text(size=12),
        axis.title = element_text(size=12),
        panel.grid = element_blank(),
        panel.background = element_rect(fill=NULL, colour='black',
                                        size=1),
        legend.position = "none") +
  xlab("Number of top genes excluded in the signature")
```

```{r}
pdf("data/figures/sup/SupplementalFigurePerformance.pdf",
    width = 10,
    height = 4.5)

plot(plot_grid(plot_grid(InPerformance_plt, OutPerformance_plt,
          align = "h",
          nrow = 1, rel_widths = c(1,1)),explegend,
          nrow = 1, rel_widths = c(1,.2)))

dev.off()
```
